1) Moved (Neutron Fwaas) Firewall related tests & `.json` mocks under `/network/firewalls/.` package. 2) Fixed minor import error.
