+++ core-test/src/main/resources/network/firewalls/firewall.json
