+++ core-test/src/main/resources/network/firewalls/firewallrule.json
