+++ core-test/src/main/resources/network/firewalls/firewallrules.json
