+++ core-test/src/main/resources/network/firewalls/firewallruleupdate.json
