+++ core-test/src/main/resources/network/firewalls/firewalls.json
